Share IOApp with IOAppPlatform traits
